survey/requirements.txt
Django==2.2.10
gunicorn==20.0.04
psycopg2-binary==2.8.6
djangorestframework==3.12.2
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0

